% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-class.R
\name{add_dim_names}
\alias{add_dim_names}
\title{Add default row and column names for a spectra object}
\usage{
add_dim_names(object)
}
\description{
Add default row and column names for a spectra object
}

