% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{add_plot_layers}
\alias{add_plot_layers}
\title{Helper to add layers to ggplot plot}
\usage{
add_plot_layers(p, plot_title = NULL, facet_cols = NULL,
  log_scale = FALSE)
}
\description{
Helper to add layers to ggplot plot
}

