% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{align_spectra}
\alias{align_spectra}
\title{Wrapper for removeBaseline and alignSpectra in MALDIquant}
\usage{
align_spectra(specmat, align_opts = list())
}
\arguments{
\item{align_opts}{A list of arguments to pass to alignSpectra in MALDIquant.}

\item{spectra_object}{A spectra object with a nonempty specmat slot.}

\item{baseline_opts}{A list of arguments to pass to removeBaseline() in
MALDIquant.}
}
\value{
aligned_peaks A matrix whose ij^th element is the value of the j^th
unique peak in the i^th unique sample.
}
\description{
Wrapper for removeBaseline and alignSpectra in MALDIquant
}

