% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectra-class.R
\name{convert_to_peaks}
\alias{convert_to_peaks}
\title{Shrink a Full Spectrum to Just Peaks}
\usage{
convert_to_peaks(physeq, peaks_list = NULL)
}
\arguments{
\item{physeq}{A phyloseqExtend object with a nonempty spectra slot.}
}
\description{
Rather than working with full spectrum reads, it is often
 convenient to work with a set of noteworthy peaks. In particular, we may
 want to first identify peaks across samples according to some peak calling
 algorithm, and then simplify the problem by setting
}

