% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectra-class.R
\name{detect_and_align_peaks}
\alias{detect_and_align_peaks}
\title{Wrapper for peak detection and alignment}
\usage{
detect_and_align_peaks(physeq, detectSpecPeaksOpts, dohClusterOpts)
}
\arguments{
\item{physeq}{A phyloseqExtend object with a nonempty spectra slot}

\item{detectSpecPeaksOpts}{A list containing options to pass into the
detectSpecPeaks function in package speaq.}

\item{dohClusterOpts}{A list containing options to pass into the
 dohCluster function in package speaq.}
}
\value{
physeq_and_peaks A list containing the following elements
   $physeq: The input physeq object, with the spectra slot modified
     so that all peaks are aligned.
   $peaks: A list of lists whose i^th element is a list of peaks for the
     i^th sample.
}
\description{
A common preprocessing step in work with spectra is peak
 alignment. Here, we wrap the peak detection, reference finding, and
 hierarchical clustering peak alignment steps from the speaq package.
}

