% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_nonzero}
\alias{filter_nonzero}
\title{Vectorized filter function in the spirit of k over a}
\usage{
filter_nonzero(X, a = 20, q = 0.9)
}
\arguments{
\item{X}{A matrix whose rows want to filter.}

\item{a}{The threshold expression value.}

\item{q}{A quantile for the number of samples above the threshold,
determining which samples we keep. So, only samples with many samples above
a will be returned.}
}
\value{
ix The indices for the rows to keep.
}
\description{
We want to keep only those rows / cols that have many values
above some threshold.
}

