% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_variance}
\alias{filter_variance}
\title{Vectorized variance filtering}
\usage{
filter_variance(X, q = 0.9)
}
\arguments{
\item{X}{A matrix whose rows want to filter.}

\item{q}{A quantile for the minimum variance of the rows we keep.}
}
\value{
X A row-variance filtered version of X.
}
\description{
We want to keep only those rows / cols that have many values
above some threshold.
}

