% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{get_annotation}
\alias{get_annotation}
\title{Helper to get annotation list from a physeq object}
\usage{
get_annotation(physeq, annotation_names = NULL)
}
\arguments{
\item{physeq}{A phyloseq object with a nonempty sample data slot}

\item{annotation_names}{A character vector giving the names of sample data
 columns to extract}
}
\value{
annotation_interaction A list containing the following elements
 $annotation: The columns of the sample_data data.frame requested from
   annotation_names
 $interaction_label: The factor vector containing all combinations of
   levels in the annotation variables.
}
\description{
Helper to get annotation list from a physeq object
}

