% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query-uniprot.R
\name{get_authors}
\alias{get_authors}
\title{Get authorship data associated with a UniProt XML}
\usage{
get_authors(uni_xml)
}
\arguments{
\item{uni_xml}{An xml_document associated with a uniprot data base,
for example, read_html(getURL("http://www.uniprot.org/uniprot/A0A015QDR1.xml"))}
}
\description{
Get authorship data associated with a UniProt XML
}

