% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{get_ppm}
\alias{get_ppm}
\title{Get ppm values associated with a spectra matrix}
\usage{
get_ppm(specmat)
}
\arguments{
\item{specmat}{A matrix whose column names will be used to derive ppms}
}
\description{
Get ppm values associated with a spectra matrix
}

