% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-class.R
\name{get_spectra_at_peaks}
\alias{get_spectra_at_peaks}
\title{Extract Spectra at Peak Positions}
\usage{
get_spectra_at_peaks(specmat, peaks_list)
}
\arguments{
\item{peaks_list}{A list whose i^th element contains the indices
of peaks in the i^th sample}
}
\value{
peaks A matrix containing spectra values filtered to samples
 and measurement indices containining peaks.
}
\description{
Extract Spectra at Peak Positions
}

