% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{get_spectra_at_peaks_zeros}
\alias{get_spectra_at_peaks_zeros}
\title{Extract Spectra at Peak Positions, with zeros for non-peaks}
\usage{
get_spectra_at_peaks_zeros(specmat, peaks_list, binary = FALSE)
}
\arguments{
\item{specmat}{A spectra matrix with samples on rows.}

\item{peaks_list}{A list whose i^th element contains the indices of peaks in
the i^th sample}

\item{binary}{Should the actual spectrum intensity be returned, or just an
indicator of whether the position was a peak?}
}
\value{
peaks A matrix containing spectra values filtered to samples and
measurement indices containining peaks, and placing zeros if a
measurement was not called a peak.
}
\description{
Extract Spectra at Peak Positions, with zeros for non-peaks
}

