% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query-uniprot.R
\name{get_uniprot_info}
\alias{get_uniprot_info}
\title{Extract information from UniProt Accession ID}
\usage{
get_uniprot_info(id)
}
\arguments{
\item{id}{The UniProt ID to extract information from.}
}
\value{
A list containing information about the specified UniProt ID.
}
\description{
Extract information from UniProt Accession ID
}
\examples{
# compare this with http://www.uniprot.org/uniprot/A0A015QDR1
get_uniprot_info("A0A015QDR1")
}

