% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectra-class.R
\name{input_peaks}
\alias{input_peaks}
\title{Fill in the peakmat Slot if a Spectrum Object}
\usage{
input_peaks(physeq, binary = FALSE, peaks_list = NULL, ...)
}
\arguments{
\item{physeq}{A phyloseqExtend object with a nonempty spectra slot.}
}
\value{
physeq The phyloseqExtend object with a peaks matrix included.
}
\description{
Rather than working with full spectrum reads, it is often
 convenient to work with a set of noteworthy peaks. In particular, we may
 want to first identify peaks across samples according to some peak calling
 algorithm, and then simplify the problem by setting
}

