% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{interpolate_duplicates}
\alias{interpolate_duplicates}
\title{Interpolate Duplicated Values}
\usage{
interpolate_duplicates(x)
}
\arguments{
\item{x}{The vector to interpolate.}
}
\value{
The interpolated vector.
}
\description{
When working with spectra, it is often necessarily to
interpolate duplicated ppm values. For example, c(1, 1, 2) should be turned
into c(1, 1.5, 2).
}
\examples{
interpolate_duplicates(c(1, 1, 2))
}

