% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processing.R
\name{merge_ctable_opts}
\alias{merge_ctable_opts}
\title{Merge default contingency table processing options}
\usage{
merge_ctable_opts(opts = list())
}
\arguments{
\item{opts}{A partially filled list specifying options to contingency table
processing. Those that are not specified will be merged according to their
defaults below. These are,
 $id_col: Is the first column of the data a nonnumeric identifier column?
  Defaults to TRUE.
 $filter_ix: A function that can be applied to the full matrix (excluding a
  potential id column), and which returns the indices of the row filtered
  version. Defaults to filter_nonzero().
 $transform: A function that can be applied to the full matrix, transforming
  its values. Defaults to log().}
}
\value{
A modified version of opts, with defaults filled in.
}
\description{
Merge default contingency table processing options
}

