% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processing.R
\name{merge_spectra_opts}
\alias{merge_spectra_opts}
\title{Merge default spectra processing options}
\usage{
merge_spectra_opts(opts = list())
}
\arguments{
\item{opts}{A partially filled list specifying options to spectra processing.
Those that are not specified will be merged according to their defaults
below. These are,
 $align_opts Arguments passed to alignSpectra() in the MALDIquant package.
  Defaults to the default options for that function. \cr
 $baseline_opts Arguments passed to the removeBaseline() function in the
  MALDIquant package. Defaults to the default options for that function. \cr
 $binary Return a sample x potential-peak indicator matrix? Or just the peaks
  intensities [with zeros, if it was not a peak]? Defaults to FALSE. \cr
 $peak_opts Arguments passed to detectPeak() in the MALDIquant package.
  Defaults to the default options for that function. \cr
 $thresh_max The maximum value for any sample, above which we discard it as
  an outlier. Defaults to Inf (don't discard anything) \cr}
}
\value{
opts The version of opts with unspecified options filled in.
}
\description{
Merge default spectra processing options
}

