% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{phyloseqExtend-class}
\alias{phyloseqExtend-class}
\title{The high-level extended phyloseq object, including new slot types}
\description{

}
\section{Slots}{

\describe{
\item{\code{spectra}}{An object of class spectra, used for storing metabolomic
  information}
}}

