% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phyloseqExtend-class.R
\docType{methods}
\name{phyloseqExtend}
\alias{phyloseqExtend}
\title{Assemble several data objects into a phyloseqExtend object.}
\usage{
phyloseqExtend(otu_table_object = NULL, sample_data_object = NULL,
  phy_tree_object = NULL, refseq_object = NULL, spectra_object = NULL)
}
\arguments{
\item{otu_table_object}{An object of class otu_table, to incorporate into
the phyloseqExtend object. Defaults to NULL.}

\item{sample_data_object}{An object of class sam_data, to incorporate into
the phyloseqExtend object. Defaults to NULL.}

\item{phy_tree_object}{An object of class phylo, to incorporate into the
phyloseqExtend object. Defaults to NULL.}

\item{refseq_object}{An object of class refseq, to incorporate into the
  phyloseqExtend object. Defaults to NULL.}

\item{spectra_object}{An object of class spectra, to incorporate into the
phyloseqExtend object. Defaults to NULL.}
}
\value{
phyloseqExtend_object An object of class phyloseqExtend, with slots
  containing each of the input data objects, and NULL for potential but
  unoccupied slots.
}
\description{
This creates a phyloseqExtend object from the multiple sources
  of data common in microbiome experiments.
}

