% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_raw_spectra}
\alias{plot_raw_spectra}
\title{Plot Raw Spectra Intensities Across All Indices}
\usage{
plot_raw_spectra(spectra_mat, method = "speaq", log_scale = FALSE,
  x_min = NULL, x_max = NULL, plot_title = NULL, ...)
}
\arguments{
\item{method}{The R package to use for the plot. Currently only supports
"speaq" or "ggplot2".}

\item{log_scale}{Should the intensities be plotted on a log scale?}

\item{x_min}{What is the minimum index to display?}

\item{x_max}{What is the maximum index to display?}

\item{plot_title}{A title to include for the figure.}
}
\description{
Before preprocessing to a collection of peaks, it may be
useful to plot the full raw spectra. This function provides an interface
to \code{speaq} and \code{ggplot2} that can be used to plot these raw
spectra.
}

