% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Plot Spectra Intensities}
\usage{
plot_spectra(physeq, method = "speaq", plot_type = "specmat",
  log_scale = FALSE, subsample_frac = 1, x_min = NULL, x_max = NULL,
  col = NULL, linetype = NULL, facet_cols = NULL, plot_title = NULL,
  alpha = 1, line_thickness = 1, ...)
}
\arguments{
\item{physeq}{A phyloseq object containing the spectrum to plot}

\item{method}{The R package to use for the plot. Currently only supports
"speaq" or "ggplot2".}

\item{log_scale}{Should the intensities be plotted on a log scale?}

\item{subsample_frac}{We will only plot ever the value at every
1 / subsample_frac indices. This can accelerate plotting in the case that}

\item{x_min}{What is the minimum index to display?}

\item{x_max}{What is the maximum index to display?}

\item{col}{A string giving a column name in sample_data(physeq) according
to which we will color the spectra.}

\item{linetype}{A string giving a column name in sample_data(physeq)
according to which we will modify the spectra linetypes.}

\item{facet_cols}{A character vector giving the column names in
sample_data(physeq) according to which to facet the plot by.}

\item{plot_title}{A title to include for the figure.}

\item{alpha}{The transparency parameter for the plots, when using ggplot2.}

\item{line_thickness}{The thickness of the spectra lines, when using ggplot2.}
}
\value{
If "ggplot2" is used, then return the ggplot2 object containing the
 plot. Otherwise does not return anything.
}
\description{
If spectra(physeq)[["peaks"]] is FALSE, then this will plot the
raw intensities of each spectrum in the sample. Otherwise, we only plot
the positions and heights of peaks for each sample.
}

