% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processing.R
\name{process_ctable}
\alias{process_ctable}
\title{Wrapper for contingency table processing}
\usage{
process_ctable(X, opts = list())
}
\arguments{
\item{X}{A matrix / data.frame / data.table with the samples as rows, and
with ids as either row names or as the first column [opts$id_col in
merge_ctable_opts()].}

\item{opts}{A list of processing options. Choices and their defaults are
detailed in merge_ctable_opts().}
}
\value{
X A processed version of the input data X.
}
\description{
Wrapper for contingency table processing
}

