% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processing.R
\name{process_spectra}
\alias{process_spectra}
\title{Wrapper for processing spectra matrices}
\usage{
process_spectra(X, opts = list())
}
\arguments{
\item{X}{A spectra matrix with samples on rows.}

\item{opts}{A potentially partially specified list of processing options. See
merge_spectra_opts() for options and defaults.}
}
\value{
A list with the followign objects
  $X_aligned The spectra matrix with peaks aligned. \cr
  $peaks_ix A list whose i^th element give the indices of peaks in the i^th
   sample. \cr
  $X_peaks A matrix whose columns are the positions where a peak was detected
   in at least one sample. \cr
  $X_peaks_zeros The same as X_peaks, but with zeros for samples where that
   given position was not identified as a peak.
}
\description{
This function 1) removes outliers, 2) removes spectra baseline,
3) aligns spectra, and 3) detects peaks, across samples in a spectrum.
}

