% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectra-class.R
\docType{methods}
\name{prune_samples,character,spectra-method}
\alias{prune_samples,character,spectra-method}
\title{Define a prune_samples methods for spectra class}
\usage{
\S4method{prune_samples}{character,spectra}(samples, x)
}
\description{
Define a prune_samples methods for spectra class
}

