% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{remove_baseline}
\alias{remove_baseline}
\title{Wrapper for removeBaseline and alignSpectra in MALDIquant}
\usage{
remove_baseline(specmat, baseline_opts = list())
}
\arguments{
\item{baseline_opts}{A list of arguments to pass to removeBaseline() in
MALDIquant.}

\item{spectra_object}{A spectra object with a nonempty specmat slot.}
}
\value{
baselined_peaks A matrix whose ij^th element is the value of the j^th
unique peak in the i^th unique sample.
}
\description{
Wrapper for removeBaseline and alignSpectra in MALDIquant
}

