% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{remove_outlier_spectra}
\alias{remove_outlier_spectra}
\title{Remove Outliers}
\usage{
remove_outlier_spectra(physeq, thresh)
}
\arguments{
\item{physeq}{A phyloseqExtend object with a nonempty spectra slot}

\item{thresh}{Any spectra with maximum value above thresh will be discarded.}
}
\value{
physeq  Version of physeq object with outlier samples removed (across
all components in the input physeq object).
}
\description{
Remove Outliers
}

