% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,phyloseqExtend-method}
\alias{show,phyloseqExtend-method}
\title{Print a phyloseqExtend object}
\usage{
\S4method{show}{phyloseqExtend}(object)
}
\description{
Print a phyloseqExtend object
}

