% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{spectra-class}
\alias{spectra-class}
\title{The S4 class for storing metabolomic information}
\description{
A matrix containing either raw spectra measurements or
   processed peaks indicators or heights.
}

