% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectra-class.R
\docType{methods}
\name{spectra}
\alias{spectra}
\title{Build or access an object of class spectra}
\usage{
spectra(object, peaks = FALSE)
}
\arguments{
\item{peaks}{A logical indicating whether the matrix columns corresponds to
 wavelengths (FALSE, for raw spectra) or just peaks (TRUE, for spectra
 that have been processed). Defaults to FALSE.}

\item{spectra_matrix}{An object of class matrix, containing the spectral
samples as rows.}
}
\value{
spectra An object of class spectra, containing the input spectra
matrix as the main data component.
}
\description{
Build or access an object of class spectra
}

