% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectra-class.R
\name{subset_samples_spectra}
\alias{subset_samples_spectra}
\title{Subset the samples of an object of class spectra}
\usage{
subset_samples_spectra(samples_ix, spectra)
}
\arguments{
\item{samples_ix}{A vector of integers giving the rows of the spectra
matrices to include.}

\item{spectra}{An object of class spectra, whose matrices we want to filter.}
}
\description{
Subset the samples of an object of class spectra
}

