% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{align_peaks}
\alias{align_peaks}
\title{Wrapper for dohCluster in speaq}
\usage{
align_peaks(spectra_object, peaks_list, ...)
}
\arguments{
\item{spectra_object}{A spectra object with a nonempty specmat slot.}

\item{peaks_list}{A list whose i^th element contains the indices
of peaks in the i^th sample}
}
\value{
aligned_peaks A matrix whose ij^th element is the value of the j^th
unique peak in the i^th unique sample.
}
\description{
Wrapper for dohCluster in speaq
}

