% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spectra-class.R
\name{get_peaks_list}
\alias{get_peaks_list}
\title{Wrapper for speaq::detectSpecPeaks}
\usage{
get_peaks_list(spectra, ...)
}
\arguments{
\item{spectra}{A spectra object containing raw spectra reads}
}
\value{
peaks_list A list whose i^th element contains the indices
of peaks in the i^th sample
}
\description{
Wrapper for speaq::detectSpecPeaks
}

