% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-utils.R
\name{input_peaks}
\alias{input_peaks}
\title{Shrink a Full Spectrum to Just Peaks}
\usage{
input_peaks(physeq, binary = FALSE, peakmat = NULL, ...)
}
\arguments{
\item{physeq}{A phyloseqExtend object with a nonempty spectra slot.}

\item{peakmat}{A matrix whose rows correspond to samples in the specmat and
whose columns index unique samples in the data.}
}
\description{
Rather than working with full spectrum reads, it is often
 convenient to work with a set of noteworthy peaks. In particular, we may
 want to first identify peaks across samples according to some peak calling
 algorithm, and then simplify the problem by setting
}

