% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_raw_spectra}
\alias{plot_raw_spectra}
\title{Plot Raw Spectra Intensities Across All Indices}
\usage{
plot_raw_spectra(physeq, method = "speaq", log_scale = FALSE,
  subsample_frac = 1, x_min = NULL, x_max = NULL, col = NULL,
  linetype = NULL, facet_cols = NULL, plot_title = NULL, ...)
}
\arguments{
\item{physeq}{A phyloseq object containing the spectrum to plot}

\item{method}{The R package to use for the plot. Currently only supports
"speaq" or "ggplot2".}

\item{log_scale}{Should the intensities be plotted on a log scale?}

\item{subsample_frac}{We will only plot ever the value at every
1 / subsample_frac indices. This can accelerate plotting in the case that
the spectrum is very long, but can lead to missed peaks.}

\item{x_min}{What is the minimum index to display?}

\item{x_max}{What is the maximum index to display?}

\item{plot_title}{A title to include for the figure.}
}
\description{
Before preprocessing to a collection of peaks, it may be
useful to plot the full raw spectra. This function provides an interface
to \code{speaq} and \code{ggplot2} that can be used to plot these raw
spectra.
}

