% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectra-class.R
\name{remove_outlier_spectra}
\alias{remove_outlier_spectra}
\title{Remove Outliers}
\usage{
remove_outlier_spectra(physeq, thresh)
}
\arguments{
\item{physeq}{A phyloseqExtend object with a nonempty spectra slot}

\item{thresh}{Any spectra with maximum value above thresh will be discarded.}
}
\value{
None. Removes the outlier samples from the spectra, sample_data and
 otu_table components of the input physeq object.
}
\description{
Remove Outliers
}

