% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{data}
\name{spectra-class}
\alias{spectra-class}
\alias{spectra_slots}
\title{The S4 class for storing metabolomic information}
\format{\preformatted{List of 4
 $ mass   : chr "numeric"
 $ specmat: chr "matrix"
 $ peakmat: chr "matrix"
 $ adjmat : chr "matrix"
}}
\usage{
spectra_slots
}
\description{
A class containing 1) a matrix with the n spectra samples
 as rows 2) a matrix with n samples as rows and with columns giving the
 union of all detected peaks, and 3) the adjacency matrix for samples
 using the coocurrence of peaks.
}
\keyword{datasets}

