% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allClasses.R
\name{spectraOrNull-class}
\alias{spectraOrNull-class}
\title{spectra S4 class with placeholder}
\description{
If a component of the phyloseq extend object is not present, we keep a NULL
   placeholder.
}
\details{

}
\keyword{internal}

